nonmem run: support Slurm submission
The new 'bbi nonmem run slurm' command follows 'bbi nonmem run sge' in
terms of going through gridSpec, attempting to align the template
directives, and testing with the same helpers.

The one deliberate deviation in behavior is the output file name.  For
SGE, standard output and standard error are redirected to the default
file name of '{run name}.o{job id}'.

There's been a request (gh-312) to clean up these .o* and .po* files
(only the former is relevant for Slurm).  Cleaning them up is probably
a bit aggressive because the .o* file contains the 'bbi run local ...'
output, which is useful for troubleshooting.  However, I suspect a
core pain point with the these files is the _changing name_ when
overwriting a model for reasons that have to do with how SVN handles
deletions.  So, don't mimic this behavior and instead use a consistent
name to hopefully avoid this issue for Slurm.

Closes #303