integration/nonmem: extract SGE checks to test helpers
bbi_sge_test.go contains two 'bbi nonmem run sge ...' tests, one for a
call without --parallel and another for a call with --parallel.  Much
of the test logic is not specific to SGE.

Extract out the core checks so that they can be reused to test the
upcoming slurm subcommand.