run-integration-tests: adjust MPIEXEC_PATH for Metworx compatibility
In the upcoming Metworx, /usr/bin/mpiexec won't exist.

We could set MPIEXEC_PATH to the default value of
/usr/local/mpich3/bin/mpiexec (or even drop the handling entirely).
The main downside is that tests could then no longer detect whether
overriding the default mpiexec path with --mpi_exec_path was
effective.

Instead set this to the _resolved_ path of the default location.  This
is a value that 1) works across Metworx versions and 2) is distinct
from the default bbi value because we can rely on it being a symlink.
(Another option would be to set it to the highest precedence mpiexec
in PATH.)

While at it, make it possible to choose a different location by
setting MPIEXEC_PATH in the calling environment.