golangci-lint: drop invalid govet.shadow setting
Based on inspecting the history of golangci-lint's
pkg/config/linters_settings.go, it doesn't look like this was a valid
option.  I think this was a typo and should have been
"check-shadowing".

check-shadowing has now been has now been removed in favor of adding
"shadow" to the govet.enable.  Enabling it gives four messages, three
of which are related to shadowing err.

Just drop the non-default setting for now.