ci: deploy site on GitHub release instead of tag push
As of the last commit, the site publishes a latest.json file with the
latest non-dev release.  bbr::use_bbi will start looking there to
determine the latest bbi release.  With that value, bbr constructs the
URL to one of the source tarballs that's included as a release asset.

A draft release is created (via GoReleaser) by the release job of the
main workflow and then finalized manually.  Deploying the site on tag
push would create a window (likely a few minutes) where the updated
site has been deployed but the release hasn't been published, causing
the use_bbi call to fail when it tries to download the source tarball.

Avoid this window by instead deploying the site when a release is
published.

Note that this workflow won't trigger for dev releases because
GoReleaser is configured to mark those as pre-releases.  And, even if
a dev tag was accidentally published as a regular release, the logic
for writing latest.json excludes dev tags.